+++ home-manager-modules/vscodium/default.nix
@@ -18,7 +18,7 @@ in {
 
   services.ollama = {
     enable = true;
-    model = "starcoder2:3b";
+    models = [ "starcoder2:3b" ];
   };
 
   programs.vscode = {
@@ -146,7 +146,7 @@ in {
       james-yu.latex-workshop
 
       # Rust
-      matklad.rust-analyzer
+      rust-lang.rust-analyzer
 
       # Spellcheck
       streetsidesoftware.code-spell-checker